backend/requirements.txt
# Blue's Book - Python Dependencies

# Core Flask Framework
Flask==2.3.3
Flask-CORS==4.0.0

# Environment and Configuration
python-dotenv==1.0.0

# Firebase SDK
firebase-admin==6.2.0

# HTTP Requests
requests==2.31.0

# Caching
redis==4.6.0

# Data Processing
pandas==2.1.1

# Date/Time Utilities
python-dateutil==2.8.2

# Image Processing
Pillow==10.0.1

# API Rate Limiting
Flask-Limiter==3.5.0

# Testing
pytest==7.4.2
pytest-flask==1.2.0
pytest-cov==4.1.0

# Development Tools
black==23.7.0
flake8==6.0.0

